Finalize auth guards

+++ net-api/src/expense/expense.service.ts
@@ -3,7 +3,6 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { AccountService } from '../account/account.service';
 import { SocketGateway } from '../socket/socket.gateway';
-import { User } from '../user/user.entity';
 import { CreateExpenseInput } from './dto/create-expense.input';
 import { UpdateExpenseInput } from './dto/update-expense.input';
 import { Expense } from './expense.entity';
@@ -17,8 +16,6 @@ export class ExpenseService {
     constructor(
         @InjectRepository(Expense)
         private readonly expenseRepository: Repository<Expense>,
-        @InjectRepository(User)
-        private readonly userRepository: Repository<User>,
         private readonly accountService: AccountService,
         private readonly socketGateway: SocketGateway,
     ) {}
@@ -30,32 +27,26 @@ export class ExpenseService {
     */
 
     async getExpenses(user_id: number) {
-        const userExists = await this.userRepository.findOneBy({ user_id });
-        if (!userExists) {
-            throw new NotFoundException(`ID ${user_id} is non-existent!`);
-        }
         return this.expenseRepository.find({
             where: { user_expense_id: user_id },
             order: { expense_id: 'ASC' },
         });
     }
 
-    async getExpense(id: number) {
-        const expense = await this.expenseRepository.findOneBy({ expense_id: id });
+    async getExpense(user_id: number, id: number) {
+        const expense = await this.expenseRepository.findOneBy({ 
+            expense_id: id,
+            user_expense_id: user_id
+         });
         if (!expense) {
             throw new NotFoundException(`Expense ID ${id} is non-existent!`);
         }
         return expense;
     }
 
-    async create(input: CreateExpenseInput) {
-        const userExists = await this.userRepository.findOneBy({ user_id: input.user_id });
-        if (!userExists) {
-            throw new NotFoundException(`User ID ${input.user_id} is non-existent!`);
-        }
-
+    async create(user_id: number, input: CreateExpenseInput) {
         const expense = this.expenseRepository.create({
-            user_expense_id: input.user_id,
+            user_expense_id: user_id,
             expense_amount: input.expense_amount,
             expense_source: input.expense_source,
             expense_date: input.expense_date,
@@ -67,20 +58,16 @@ export class ExpenseService {
         // Negative sign because adjustbalance is called with .increment()
         await this.accountService.adjustBalance(input.account_expense_id, -input.expense_amount);
 
-        this.socketGateway.broadcast('expense.changed', { user_id: input.user_id });
-        return this.getExpense(saved.expense_id);
+        this.socketGateway.broadcast('expense.changed', { user_id });
+        return saved;
     }
 
-    async update(id: number, input: UpdateExpenseInput) {
-        const oldExpense = await this.getExpense(id);
+    async update(user_id: number, id: number, input: UpdateExpenseInput) {
+        const oldExpense = await this.getExpense(id, user_id);
 
-        await this.expenseRepository.update(id, {
-            expense_amount: input.expense_amount,
-            expense_source: input.expense_source,
-            expense_date: input.expense_date,
-            category_expense_id: input.category_expense_id,
-            account_expense_id: input.account_expense_id,
-        });
+        await this.expenseRepository.update({ expense_id: id, user_expense_id: user_id}, input);
+
+        const updatedExpense = await this.getExpense(id, user_id);
 
         // Removes previous amount affecting account to prevent stacking
         await this.accountService.adjustBalance(
@@ -90,23 +77,26 @@ export class ExpenseService {
 
         // Adds the new amount
         await this.accountService.adjustBalance(
-            input.account_expense_id,
+            updatedExpense.account_expense_id,
             -Number(input.expense_amount),
         );
 
-        this.socketGateway.broadcast('expense.changed', { user_id: oldExpense.user_expense_id });
-        return this.getExpense(id);
+        this.socketGateway.broadcast('expense.changed', { user_id});
+        return updatedExpense;
     }
 
-    async remove(id: number) {
-        const expense = await this.getExpense(id);
-
+    async remove(user_id: number, id: number) {
+        const current = await this.getExpense(user_id, id);
+        if (!current) {
+            throw new NotFoundException(`Account ID ${id} is non-existent`);
+        }
+        // Refunds the amount back
         await this.accountService.adjustBalance(
-            expense.account_expense_id,
-            Number(expense.expense_amount),
+            current.account_expense_id,
+            Number(current.expense_amount),
         );
 
         await this.expenseRepository.delete(id);
-        this.socketGateway.broadcast('expense.changed', { user_id: expense.user_expense_id });
+        this.socketGateway.broadcast('expense.changed', { user_id });
     }
 }
